Refactor memory fill operations and enhance texture handling in GPU tests
- Updated memory fill base address to 24 bits across various test scripts to accommodate larger address space.
- Modified texture upload logic in ver_016.py to utilize a 64x64 checker texture with 16x16 squares, implementing bilinear filtering.
- Adjusted Z-buffer clearing operations in multiple test scripts to reflect new memory addressing.
- Enhanced triangle emission logic in ver_016.py to support rendering of road strips instead of a single quad, improving depth interpolation.
- Added a new helper function for emitting block-aligned checker textures.
- Updated register file testbench to handle new memory fill layout, ensuring compatibility with the updated GPU architecture.

+++ spi_gpu/dt/gs-twin/src/reg.rs
@@ -105,8 +105,8 @@ pub enum GpuAction {
 
     /// Hardware memory fill (REQ-005.08).
     MemFill {
-        /// Fill-base register field (byte address = `base << 9`).
-        base: u16,
+        /// Fill-base word address (byte address = `base * 2`).
+        base: u32,
         /// 16-bit fill value.
         value: u16,
         /// Number of 16-bit words to fill.
